Refactor sync tests to parameterize tests. Allows tests to be used with various versions of WebDriver, e.g. ChromeDriver and geckodriver. (#152)
* Refactor tests to allow passing of WebDriver creation function. Allows same test to be used for Firefox and Chrome.

* Ran dartfmt over test dir.
