+++ test/sync/command_event.dart
@@ -21,14 +21,14 @@ import 'package:webdriver/sync_core.dart';
 
 import 'sync_io_config.dart' as config;
 
-void main() {
+void runTests(config.createTestDriver createTestDriver) {
   group('CommandEvent', () {
     WebDriver driver;
 
     var events = <WebDriverCommandEvent>[];
 
     setUp(() {
-      driver = config.createTestDriver();
+      driver = createTestDriver();
       driver.addEventListener(events.add);
       driver.get(config.testPagePath);
     });
