+++ test/sync/navigation.dart
@@ -13,19 +13,20 @@
 // limitations under the License.
 
 @TestOn("vm")
-library webdriver.options_test;
+library webdriver.navigation_test;
 
 import 'package:test/test.dart';
 import 'package:webdriver/sync_core.dart';
 
 import 'sync_io_config.dart' as config;
 
-void main() {
-  group('TimeOuts', () {
+void runTests(config.createTestDriver createTestDriver) {
+  group('Navigation', () {
     WebDriver driver;
 
     setUp(() {
-      driver = config.createTestDriver();
+      driver = createTestDriver();
+      driver.get(config.testPagePath);
     });
 
     tearDown(() {
@@ -35,11 +36,15 @@ void main() {
       driver = null;
     });
 
-    // TODO(DrMarcII): Figure out how to tell if timeouts are correctly set
-    test('set all timeouts', () {
-      driver.timeouts.setScriptTimeout(new Duration(seconds: 5));
-      driver.timeouts.setImplicitTimeout(new Duration(seconds: 1));
-      driver.timeouts.setPageLoadTimeout(new Duration(seconds: 10));
+    test('refresh', () {
+      var element = driver.findElement(const By.tagName('button'));
+      driver.navigate.refresh();
+      try {
+        element.name;
+      } on StaleElementReferenceException {
+        return true;
+      }
+      return 'expected StaleElementReferenceException';
     });
   }, timeout: new Timeout(new Duration(minutes: 1)));
 }
